services/graph-service/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
kafka-python==2.0.2
neo4j==5.15.0
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0